requirements.txt
asgiref==3.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap4==2025.6
cryptography==46.0.2
Django==5.2.7
django-allauth==65.12.0
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
idna==3.11
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
review==1510
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
django-cors-headers==4.3.1
